Rename data to uint16 texture
